Fix casting int to short

+++ src/main/java/com/faforever/api/deployment/LegacyFeaturedModDeploymentTask.java
@@ -97,7 +97,7 @@ public class LegacyFeaturedModDeploymentTask implements Runnable {
     String branch = featuredMod.getGitBranch();
     boolean allowOverride = Optional.ofNullable(featuredMod.isAllowOverride()).orElse(false);
     String modFilesExtension = featuredMod.getFileExtension();
-    Map<String, Short> fileIds = featuredModService.getFileIds(modName);
+    Map<String, Integer> fileIds = featuredModService.getFileIds(modName);
 
     log.info("Starting deployment of '{}' from '{}', branch '{}', allowOverride '{}', modFilesExtension '{}'",
       modName, repositoryUrl, branch, allowOverride, modFilesExtension);
@@ -147,9 +147,9 @@ public class LegacyFeaturedModDeploymentTask implements Runnable {
    * featured mod.
    */
   @SneakyThrows
-  private Optional<StagedFile> createPatchedExe(short version, Map<String, Short> fileIds, Path targetFolder) {
+  private Optional<StagedFile> createPatchedExe(short version, Map<String, Integer> fileIds, Path targetFolder) {
     String clientFileName = "ForgedAlliance.exe";
-    Short fileId = fileIds.get(clientFileName);
+    Integer fileId = fileIds.get(clientFileName);
     if (fileId == null) {
       log.debug("Skipping '{}' because there's no file ID available", clientFileName);
       return Optional.empty();
@@ -210,7 +210,7 @@ public class LegacyFeaturedModDeploymentTask implements Runnable {
    */
   @SneakyThrows
   @SuppressWarnings("unchecked")
-  private List<StagedFile> packageFiles(Path repositoryDirectory, short version, Map<String, Short> fileIds, Path targetFolder) {
+  private List<StagedFile> packageFiles(Path repositoryDirectory, short version, Map<String, Integer> fileIds, Path targetFolder) {
     updateStatus("Packaging files");
     try (Stream<Path> stream = Files.list(repositoryDirectory)) {
       return stream
@@ -248,14 +248,14 @@ public class LegacyFeaturedModDeploymentTask implements Runnable {
    * content of the directory. If no file ID is available, an empty optional is returned.
    */
   @SneakyThrows
-  private Optional<StagedFile> packDirectory(Path directory, Short version, Path targetFolder, Map<String, Short> fileIds) {
+  private Optional<StagedFile> packDirectory(Path directory, Short version, Path targetFolder, Map<String, Integer> fileIds) {
     String directoryName = directory.getFileName().toString();
     Path targetNxtFile = targetFolder.resolve(String.format("%s.%d.%s", directoryName, version, featuredMod.getFileExtension()));
     Path tmpNxtFile = toTmpFile(targetNxtFile);
 
     // E.g. "effects.nx2"
     String clientFileName = String.format("%s.%s", directoryName, featuredMod.getFileExtension());
-    Short fileId = fileIds.get(clientFileName);
+    Integer fileId = fileIds.get(clientFileName);
     if (fileId == null) {
       log.debug("Skipping folder '{}' because there's no file ID available", directoryName);
       return Optional.empty();
@@ -360,7 +360,7 @@ public class LegacyFeaturedModDeploymentTask implements Runnable {
     /**
      * ID of the file as stored in the database.
      */
-    short fileId,
+    int fileId,
     /**
      * The staged file, already in the correct location, that is ready to be renamed.
      */
